Norm sorted
The problem with the norm was due to the strange order of final indices after performing einsum. Therefore, everywhere where I use this function, I specify the order of indices that are expected.However, the precision for length of running is still quite low, therefore I try to calculate and implement higher order terms.
git commit -m